Updating SSLClients.hpp and SSLClients.cpp header and source files.
To remove superfluous data structures.

+++ Source/Utilities.cpp
@@ -819,7 +819,7 @@ namespace DiscordCoreAPI {
 		}
 		newBuffer[1] = encodedBytes;
 		newBuffer[2] = 0;
-		this->writeString(newBuffer, 1 + 2 + encodedBytes);
+		this->writeString(newBuffer, static_cast<size_t>(1 + 2 + encodedBytes));
 	}
 
 	void Jsonifier::appendNewFloatExt(const double FloatValue) {
